dom/indexedDB/test/test_filehandle_serialization.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    const READ_WRITE = "readwrite";

    const databaseInfo = [
      { name: window.location.pathname + "1" },
      { name: window.location.pathname + "2" }
    ];

    const objectStoreName = "Blobs";

    let databases = [];
    for (let info of databaseInfo) {
      let request = indexedDB.open(info.name, 1);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      db.createObjectStore(objectStoreName, { });

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      databases.push(db);
    }

    let db1 = databases[0];

    let request = db1.createMutableFile("random.bin", "binary/random");
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let mutableFile = event.target.result;
    is(getFileId(mutableFile), 1, "Correct file id");
    is(mutableFile.name, "random.bin", "Correct name");
    is(mutableFile.type, "binary/random", "Correct type");

    let trans = db1.transaction([objectStoreName], READ_WRITE);
    let objectStore = trans.objectStore(objectStoreName);

    request = objectStore.add(mutableFile, 42);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    request = objectStore.get(42);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let result = event.target.result;
    is(getFileId(result), 1, "Correct file id");
    is(result.name, mutableFile.name, "Correct name");
    is(result.type, mutableFile.type, "Correct type");

    let db2 = databases[1];

    trans = db2.transaction([objectStoreName], READ_WRITE);
    objectStore = trans.objectStore(objectStoreName);

    try {
      objectStore.add(mutableFile, 42);
      ok(false, "Should have thrown!");
    }
    catch (e) {
      ok(e instanceof DOMException, "Got exception.");
      is(e.name, "DataCloneError", "Good error.");
      is(e.code, DOMException.DATA_CLONE_ERR, "Good error code.");
    }

    finishTest();
  }
  </script>
  <script type="text/javascript" src="file.js"></script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
